+++ .github/workflows/native.yml
@@ -4,13 +4,13 @@ on: [ push, workflow_dispatch ]
 
 env:
   BUILD_TYPE: Debug
-  PROJECT_SUBDIR: sgbmandroid/src/main/cpp
+  PROJECT_SUBDIR: sgbmdepth/src/main/cpp
   BUILD_SUBDIR: build
 
 defaults:
   run:
     shell: bash
-    working-directory: sgbmandroid/src/main/cpp  # Must be the same as env.PROJECT_SUBDIR
+    working-directory: sgbmdepth/src/main/cpp  # Must be the same as env.PROJECT_SUBDIR
 
 jobs:
   build:
@@ -25,7 +25,7 @@ jobs:
         uses: actions/cache@v3
         with:
           path: ${{github.workspace}}/${{env.PROJECT_SUBDIR}}/${{env.BUILD_SUBDIR}}
-          key: ${{runner.os}}-build-cache-v4
+          key: ${{runner.os}}-build-cache
 
       - name: Prepare OpenCV
         run: cmake -P BuildOpenCV.cmake
@@ -57,7 +57,7 @@ jobs:
           path: |
             ~/.cmodule
             ${{github.workspace}}/${{env.PROJECT_SUBDIR}}/${{env.BUILD_SUBDIR}}
-          key: ${{runner.os}}-build-cache-bridge-v2
+          key: ${{runner.os}}-build-cache-bridge
 
       - name: Prepare OpenCV
         run: cmake -P BuildOpenCV.cmake
@@ -81,7 +81,7 @@ jobs:
         uses: actions/cache@v3
         with:
           path: ${{github.workspace}}/${{env.PROJECT_SUBDIR}}/${{env.BUILD_SUBDIR}}
-          key: ${{runner.os}}-build-cache-v4
+          key: ${{runner.os}}-build-cache
 
       - name: Configure linters
         run: |
