+++ app/build.gradle
@@ -7,10 +7,10 @@ plugins {
 android {
     compileSdk 31
 
-    namespace 'me.timpushkin.sgbmandroidapp'
+    namespace 'me.timpushkin.sgbmdepthapp'
 
     defaultConfig {
-        applicationId 'me.timpushkin.sgbmandroidapp'
+        applicationId 'me.timpushkin.sgbmdepthapp'
         minSdk 21
         targetSdk 31
         versionCode 1
@@ -57,7 +57,7 @@ dependencies {
     implementation 'com.squareup.okio:okio:3.1.0'
     implementation 'com.google.modernstorage:modernstorage-storage:1.0.0-alpha06'
 
-    implementation project(':sgbmandroid')
+    implementation project(':sgbmdepth')
 }
 
 ktlint {
