+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">SgbmAndroid</string>
+    <string name="app_name">SgbmDepth</string>
 </resources>
