+++ sgbmdepth/src/main/AndroidManifest.xml
