+++ sgbmdepth/src/main/cpp/.clang-tidy
