+++ sgbmdepth/src/main/cpp/.gitignore
