+++ sgbmdepth/src/main/cpp/CPPLINT.cfg
