+++ sgbmdepth/src/main/cpp/external/log-lib/CMakeLists.txt
